Cleanup imports, and code structure.
