+++ examples/tests/test_device_viewer_widget.py
@@ -4,7 +4,7 @@ import numpy as np
 from traits.trait_errors import TraitError
 
 from device_viewer.models.electrodes import Electrodes
-from device_viewer.utils.dmf_utils import channels_to_svg
+from device_viewer.utils.dmf_utils_helpers import channels_to_svg
 from .common import TEST_PATH
 from pathlib import Path
 
@@ -150,7 +150,7 @@ def test_electrodes_states_map(valid_electrodes_model_from_svg):
 def test_save_svg_file_init_scale_metadata(valid_electrodes_model_from_svg):
     # save file that initially does not have a scale, with a scale
 
-    from device_viewer.utils.dmf_utils import channels_to_svg
+    from device_viewer.utils.dmf_utils_helpers import channels_to_svg
 
     new_filename = Path(TEST_PATH) / "test_svg_model_save_init_scale.svg"
     new_pixel_scale = 0.55
@@ -168,7 +168,7 @@ def test_save_svg_file_init_scale_metadata(valid_electrodes_model_from_svg):
 def test_save_svg_file_edit_scale_metadata(valid_electrodes_model_from_svg_with_scale):
     # save file that initially does have a scale, with a new scale
 
-    from device_viewer.utils.dmf_utils import channels_to_svg
+    from device_viewer.utils.dmf_utils_helpers import channels_to_svg
 
     # check that the loaded file is as expected
     assert valid_electrodes_model_from_svg_with_scale.svg_model.area_scale == 0.34
